+++ build_modules/lib/src/kernel_builder.dart
@@ -32,7 +32,7 @@ class KernelSummaryBuilder implements Builder {
   @override
   Future build(BuildStep buildStep) async {
     var module = new Module.fromJson(
-        JSON.decode(await buildStep.readAsString(buildStep.inputId))
+        json.decode(await buildStep.readAsString(buildStep.inputId))
             as Map<String, dynamic>);
     try {
       await createKernelSummary(module, buildStep);